Feat: RecordController 구체 클래스로 변경
